refactor: feature folder name changed to domains
