+++ src/app/(store)/list/[[...params]]/page.tsx
@@ -13,10 +13,10 @@ import Button from "@/components/UI/button";
 import DropDownList from "@/components/UI/dropDown";
 import LineList from "@/components/UI/lineList";
 import { sortDropdownData } from "@/data/uiElementsData";
-import { ProductListSkeleton } from "@/features/store/productList/components";
-import { DEFAULT_FILTERS, SORT_DATA } from "@/features/store/productList/constants";
-import { TFilterBrands, TFilters, TListItem } from "@/features/store/productList/types";
-import { getFiltersFromProductList } from "@/features/store/productList/utils";
+import { ProductListSkeleton } from "@/domains/store/productList/components";
+import { DEFAULT_FILTERS, SORT_DATA } from "@/domains/store/productList/constants";
+import { TFilterBrands, TFilters, TListItem } from "@/domains/store/productList/types";
+import { getFiltersFromProductList } from "@/domains/store/productList/utils";
 import { IMAGE_BASE_URL } from "@/shared/constants/store";
 import { cn } from "@/shared/utils/styling";
 import { TPageStatus } from "@/types/list";
