+++ src/app/admin/brands/page.tsx
@@ -5,7 +5,7 @@ import { addBrand, deleteBrand, getAllBrands, updateBrand } from "@/actions/bran
 import Button from "@/components/UI/button";
 import Input from "@/components/UI/input";
 import Popup from "@/components/UI/popup";
-import { TBrand } from "@/types/product";
+import { TBrand } from "@/shared/types";
 
 let selectedBrandID = "";
 const Brand = () => {
