+++ src/components/store/listPage/filters/index.tsx
@@ -5,9 +5,9 @@ import Button from "@/components/UI/button";
 import CheckBox from "@/components/UI/checkBox";
 import PriceSlider from "@/components/UI/priceSlider";
 import { SK_Box } from "@/components/UI/skeleton";
+import { TFilters } from "@/domains/store/productList/types";
 import { cn } from "@/shared/utils/styling";
 import { TPageStatus } from "@/types/list";
-import { TFilters } from "@/types/product";
 
 type TProps = {
   showFilters: boolean;
