+++ src/domains/product/constants/index.ts
