+++ src/domains/product/types/dealCards.ts
