+++ src/domains/product/types/index.ts
