+++ src/domains/store/homePage/components/categories/index.tsx
