+++ src/domains/store/homePage/components/companyLogo/CompanyLogo.tsx
