+++ src/domains/store/homePage/components/companyLogo/index.tsx
