+++ src/domains/store/homePage/components/slider/index.tsx
