+++ src/domains/store/homePage/components/wideCardRow/index.tsx
