+++ src/domains/store/homePage/constants/index.ts
