+++ src/domains/store/homePage/constants/wideCards.ts
