+++ src/domains/store/homePage/types/collection.ts
