+++ src/domains/store/homePage/types/index.ts
