+++ src/domains/store/homePage/types/wideCard.ts
