+++ src/domains/store/productList/types/index.ts
